# Conflicts:
#	Application/Starball/View/Public/header.html
#	Public/starball/css/test.css

+++ Application/Common/Logic/ItemPriceLogic.class.php
@@ -57,6 +57,9 @@
 
 		public function deleteItemPricesByItemId($itemId) {
 			$map["itemId"] = $itemId;
+			if ($this->where($map)->count() == 0) {
+				return true;
+			}
 			if ($this->where($map)->delete() !== false) {
 				return true;
 			} else {
